Add verify-suite script
