+++ package.json
@@ -38,6 +38,9 @@
     "deploy-suite:ethereum:mainnet": "npx hardhat run --network mainnet scripts/deploy-suite.js >> logs/mainnet.deploy.contracts.txt",
     "deploy-suite:polygon:mumbai": "npx hardhat run --network mumbai scripts/deploy-suite.js >> logs/mumbai.deploy.contracts.txt",
     "deploy-suite:polygon:mainnet": "npx hardhat run --network polygon scripts/deploy-suite.js >> logs/polygon.deploy.contracts.txt",
+    "verify-suite:ethereum:mainnet": "npx hardhat run --network mainnet scripts/verify-suite.js >> logs/mainnet.verify.contracts.txt",
+    "verify-suite:polygon:mumbai": "npx hardhat run --network mumbai scripts/verify-suite.js >> logs/polygon.verify.contracts.txt",
+    "verify-suite:polygon:mainnet": "npx hardhat run --network polygon scripts/verify-suite.js >> logs/polygon.verify.contracts.txt",
     "deploy-mocks:hardhat": "npx hardhat clean && npx hardhat compile && npx hardhat deploy-mock-nft-auth --network hardhat",
     "deploy-mocks:local": "npx hardhat clean && npx hardhat compile && npx hardhat deploy-mock-nft-auth --network localhost",
     "deploy-mocks:test": "npx hardhat clean && npx hardhat compile && npx hardhat deploy-mock-nft-auth --network test >> logs/test.deploy.mocks.txt",
